requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
celery==5.4.0
redis==5.2.1
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
drf-spectacular==0.28.0
django-filter==24.3
alpaca-trade-api==3.2.0
websockets==14.1
dj-database-url==2.2.0
TA-Lib==0.4.32
